fix: use of long, not size_t when using erase.

+++ src/hash_set_coarse_grained.h
@@ -37,7 +37,7 @@ class HashSetCoarseGrained : public HashSetBase<T> {
 
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (size_t i = 0; i < bucket.size(); i++) {
+    for (long i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
         bucket.erase(bucket.begin() + i);
         return true;
