+++ src/hash_set_refinable.h
@@ -53,7 +53,7 @@ class HashSetRefinable : public HashSetBase<T> {
 
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (size_t i = 0; i < bucket.size(); i++) {
+    for (long i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
         bucket.erase(bucket.begin() + i);
         return true;
