+++ src/hash_set_sequential.h
@@ -31,7 +31,7 @@ class HashSetSequential : public HashSetBase<T> {
   bool Remove(T elem) final {
     bucket_t<T>& bucket = GetBucket(elem);
 
-    for (size_t i = 0; i < bucket.size(); i++) {
+    for (long i = 0; i < bucket.size(); i++) {
       if (bucket[i] == elem) {
         bucket.erase(bucket.begin() + i);
         return true;
